[avl tree] move Node class to node header
- move Node class to node header
